requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pillow
cloudinary
